--- frontend/src/components/Stepper.css
@@ -1,172 +1,0 @@
-.outer-container {
-  display: flex;
-  min-height: 100%;
-  flex: 1 1 0%;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 1rem;
-}
-
-@media (min-width: 640px) {
-  .outer-container {
-    aspect-ratio: 4 / 3;
-  }
-}
-
-@media (min-width: 768px) {
-  .outer-container {
-    aspect-ratio: 2 / 1;
-  }
-}
-
-.step-circle-container {
-  margin-left: auto;
-  margin-right: auto;
-  width: 100%;
-  max-width: 28rem;
-  border-radius: 2rem;
-  box-shadow: 0 28px 48px rgba(255, 114, 239, 0.18),
-    0 8px 20px rgba(155, 94, 255, 0.12);
-  background: rgba(22, 19, 36, 0.92);
-  border: 1px solid rgba(255, 114, 239, 0.25);
-  backdrop-filter: blur(18px);
-}
-
-.step-indicator-row {
-  display: flex;
-  width: 100%;
-  align-items: center;
-  padding: 2rem;
-  gap: 0.5rem;
-}
-
-.step-content-default {
-  position: relative;
-  overflow: hidden;
-  padding-bottom: 1rem;
-  color: #f8f8fa;
-}
-
-.step-default {
-  padding-left: 2rem;
-  padding-right: 2rem;
-}
-
-.footer-container {
-  padding-left: 2rem;
-  padding-right: 2rem;
-  padding-bottom: 2rem;
-}
-
-.footer-nav {
-  margin-top: 2.5rem;
-  display: flex;
-}
-
-.footer-nav.spread {
-  justify-content: space-between;
-}
-
-.footer-nav.end {
-  justify-content: flex-end;
-}
-
-.back-button {
-  transition: all 350ms;
-  border-radius: 0.25rem;
-  padding: 0.35rem 0.75rem;
-  color: #b3accb;
-  cursor: pointer;
-  background: rgba(255, 255, 255, 0.04);
-  border: 1px solid rgba(255, 114, 239, 0.15);
-  font-weight: 500;
-}
-
-.back-button:hover {
-  color: #fff;
-  border-color: rgba(255, 114, 239, 0.35);
-}
-
-.back-button.inactive {
-  pointer-events: none;
-  opacity: 0.4;
-  color: #615b79;
-}
-
-.next-button {
-  transition: all 350ms;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 9999px;
-  background-image: linear-gradient(135deg, #ff72ef 0%, #9b5eff 100%);
-  color: #fff;
-  font-weight: 600;
-  letter-spacing: -0.01em;
-  padding: 0.5rem 1.7rem;
-  cursor: pointer;
-  border: none;
-  box-shadow: 0 16px 30px rgba(255, 114, 239, 0.28);
-}
-
-.next-button:hover {
-  box-shadow: 0 20px 36px rgba(155, 94, 255, 0.32);
-  transform: translateY(-1px);
-}
-
-.next-button:active {
-  transform: translateY(1px);
-}
-
-.step-indicator {
-  position: relative;
-  outline: none;
-}
-
-.step-indicator-inner {
-  display: flex;
-  height: 3rem;
-  width: 3rem;
-  align-items: center;
-  justify-content: center;
-  border-radius: 9999px;
-  font-weight: 600;
-  box-shadow: inset 0 0 16px rgba(255, 114, 239, 0.12);
-}
-
-.active-dot {
-  height: 0.85rem;
-  width: 0.85rem;
-  border-radius: 9999px;
-  background-color: #fff;
-}
-
-.step-number {
-  font-size: 0.9rem;
-  color: inherit;
-}
-
-.step-connector {
-  position: relative;
-  margin-left: 0.75rem;
-  margin-right: 0.75rem;
-  height: 0.125rem;
-  flex: 1;
-  overflow: hidden;
-  border-radius: 0.25rem;
-  background-color: rgba(82, 82, 91, 0.4);
-}
-
-.step-connector-inner {
-  position: absolute;
-  left: 0;
-  top: 0;
-  height: 100%;
-}
-
-.check-icon {
-  height: 1.1rem;
-  width: 1.1rem;
-  color: #fff;
-}
