+++ frontend/src/styles/common.css
@@ -1,0 +1,145 @@
+/* ===================================
+   공통 스타일 (Common Styles)
+   =================================== */
+
+/* 공통 그라데이션 */
+.gradient-primary {
+  background: linear-gradient(90deg, #6dd5fa 0%, #c4b5fd 50%, #fca5f1 100%);
+}
+
+.gradient-text {
+  background: linear-gradient(90deg, #6dd5fa 0%, #c4b5fd 50%, #fca5f1 100%);
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+}
+
+.gradient-border {
+  border-image: linear-gradient(90deg, #6dd5fa 0%, #c4b5fd 50%, #fca5f1 100%);
+  border-image-slice: 1;
+}
+
+/* 공통 폰트 패밀리 */
+:root {
+  --font-primary: "Noto Sans KR", -apple-system, BlinkMacSystemFont, "Segoe UI",
+    sans-serif;
+}
+
+/* 공통 색상 변수 */
+:root {
+  --color-sky-blue: #6dd5fa;
+  --color-purple: #c4b5fd;
+  --color-pink: #fca5f1;
+  --color-error: #f44336;
+  --color-warning: #ffc107;
+  --color-text-primary: #1a1a1a;
+  --color-text-secondary: #666;
+  --color-text-muted: #999;
+  --color-bg-primary: #fafafa;
+  --color-bg-secondary: #fff;
+  --color-border: #e0e0e0;
+}
+
+/* 공통 버튼 스타일 */
+.btn-primary {
+  padding: 12px 28px;
+  border-radius: 999px;
+  border: 1px solid #000;
+  background: #000;
+  color: #fff;
+  font-weight: 700;
+  font-size: 14px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.btn-primary:hover {
+  background: #fff;
+  color: #000;
+}
+
+.btn-secondary {
+  padding: 8px 22px;
+  border: 1px solid #000;
+  border-radius: 999px;
+  background: #fff;
+  color: #000;
+  font-size: 15px;
+  font-weight: 600;
+  cursor: pointer;
+  transition: background 0.2s ease, color 0.2s ease;
+}
+
+.btn-secondary:hover {
+  background: #000;
+  color: #fff;
+}
+
+/* 공통 카드 스타일 */
+.card {
+  background: #fff;
+  border-radius: 12px;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.06);
+  transition: all 0.2s ease;
+}
+
+.card:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 6px 12px rgba(196, 181, 253, 0.3);
+}
+
+/* 공통 입력 필드 스타일 */
+.input-field {
+  width: 100%;
+  padding: 12px 16px;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  font-size: 14px;
+  outline: none;
+  transition: all 0.2s ease;
+}
+
+.input-field:focus {
+  border-color: #c4b5fd;
+  box-shadow: 0 0 0 3px rgba(196, 181, 253, 0.1);
+}
+
+/* 공통 애니메이션 */
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes slideIn {
+  from {
+    opacity: 0;
+    transform: translateY(10px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+/* 공통 스크롤바 스타일 */
+.custom-scrollbar::-webkit-scrollbar {
+  width: 8px;
+  height: 8px;
+}
+
+.custom-scrollbar::-webkit-scrollbar-track {
+  background: transparent;
+}
+
+.custom-scrollbar::-webkit-scrollbar-thumb {
+  background: #d0d0d0;
+  border-radius: 4px;
+}
+
+.custom-scrollbar::-webkit-scrollbar-thumb:hover {
+  background: #b0b0b0;
+}
